+++ jobbergate-api/CHANGELOG.md
@@ -6,6 +6,7 @@ This file keeps track of all notable changes to jobbergate-api
 
 - Added constraints to prevent long strings from being inserted into the database
 - Added constraints to prevent negative id numbers from being inserted into the database
+- Added constraints to limit the size on the uploaded files
 
 ## 4.2.0a6 -- 2023-12-12
 ## 4.2.0a5 -- 2023-12-12